requirements.txt
aiohttp==3.6.2
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.10.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
blinker==1.4
boto3==1.15.16
botocore==1.18.16
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
chatbotAI==0.2.2.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.4.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.5
djangorestframework==3.12.1
djongo==1.3.3
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docx2txt==0.8
duplicity==0.8.12.0
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
httplib2==0.14.0
idna==2.8
IMAPClient==2.1.0
importlib-metadata==1.5.0
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
multidict==4.7.6
netifaces==0.10.4
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
paho-mqtt==1.5.1
pandas==1.1.3
paramiko==2.6.0
pdf2image==1.14.0
pdfminer.six==20181108
pexpect==4.6.0
Pillow==7.0.0
pkg==0.2
protobuf==3.6.1
pycairo==1.16.2
pycryptodome==3.9.8
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
PyPDF2==1.26.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-pptx==0.6.18
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2020.10.15
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
s3transfer==0.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.20
sqlparse==0.2.4
systemd-python==234
textract==1.6.3
texttable==1.6.2
toml==0.10.1
tqdm==4.50.2
tzlocal==1.5.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.3.7
xmltodict==0.12.0
yarl==1.5.1
zipp==1.0.0
